third_party/WebKit/LayoutTests/animations/timing/timing-functions-neutral-keyframe.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
  @keyframes anim {
    from { animation-timing-function: linear; }
    to { left: 500px; }
  }

  #target {
    animation: anim 10s -1s paused ease-in-out;
    left: 0px;
    position: absolute;
  }
</style>
<div id="target"></div>
<script>
  test(function() {
    assert_equals(parseInt(getComputedStyle(target).left), 50, 'left offset'); // Should be linearly interpolated
  }, "Check that explicitly specified timing functions in neutral keyframes override default");
</script>
